chore: improve Transition exports
